+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/synthesis/SynthesisCreateScreen.java
@@ -221,7 +221,7 @@ public class SynthesisCreateScreen extends MenuFilterable {
 	@Override
 	protected void renderSelectedData(GuiGraphics gui, int mouseX, int mouseY, float partialTicks) {
 		PoseStack matrixStack = gui.pose();
-		float tooltipPosX = width * 0.3333F;
+		float tooltipPosX = bottomRightBar.getPosX() + 8;
 		float tooltipPosY = height * 0.8F;
 
 		float iconPosX = boxRT.getX();