+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/synthesis/SynthesisForgeScreen.java
@@ -231,7 +231,7 @@ public class SynthesisForgeScreen extends MenuFilterable {
 	@Override
 	protected void renderSelectedData(GuiGraphics gui, int mouseX, int mouseY, float partialTicks) {
 		PoseStack matrixStack = gui.pose();
-		float tooltipPosX = width * 0.3333F;
+		float tooltipPosX = bottomRightBar.getPosX() + 8;
 		float tooltipPosY = height * 0.8F;
 
 		float iconPosX = boxR.getX();