+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Added:
+* Common config option for gummi fuel consumption factor, defaulting to half of what it was before.
+* New achievement for when obtaining a winner stick.
+
+### Changed:
+* 
+
 ### Fixed:
+* Magic Target Block didn't have a recipe, it is now crafted with a Target block and a Redstone block.
 * Culled magnet blox trails based on distance, increasing performance when they are far from the player.
 * Made Full MP Blast stackable with diminishing returns since the description mentioned it.
+* Synthesis and Keyblade Forge screens item description being misaligned.